drivers/eeprom: at24: correct and quicker writes to eeprom.
The device is now polled for write completion. All the datasheets I looked
at had a max of 10ms for eeprom write time.

Writes across page boundaries are now handled correctly; the device
automatically wraps the write to page boundary, so we don't write more
than is remaining in the page.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>