vpr200: use flash header to init SDRAM, not low level init
The vpr200 always uses internal boot mode so the flash header can be used
to init the SD ram. The memory init in board_init_lowlevel is redundant
and can be removed.
(Although, to be honest I'm still not sure what the Load Mode Register
stuff is all about -- that was slightly different.)

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>