vpr200: enable charge shunt on startup.
This enables the charge shunt on startup to assist booting with a low
battery when the charger is attached.
For some reason the charger supply drops out when the battery is low
 - even when the pmic has the charger present, power limiting is disabled
and charge current threshold is well above current usage.
Enabling the charge shunt ensures sufficient supply to the device.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>