vpr200_tester: add "tsinputs_test" command to test TOUCHnn inputs
This adds a command to test the TOUCHnn inputs.
Each input is sampled 8 times and then averaged. The average value needs
to be within 15% of a reference value in order to be successful.

Required hardware is a string of 8 equal value resistors between a 3.15V
and GND, and then taking the 4 lowest taps to channel 12..15.
(Eight resistors are used so that the highest used value is below the 2.4V
input buffer).

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>